rework colour-select dropdown
